Change a bound check from `==` to `>=` to allow the compiler to reason more.
Letting the compiler know that the case where the index is further off the
end of the slice than just past the end should allow for a bound check to
be eliminated on the line immediately below. Benchmarking suggests this
might have a good effect and at least doesn't have a bad effect, but I didn't
check the generated assembly.