Use packed_simd::IntoBits instead of FromBits
As of rustc 1.59.0, all the simd intrinsics in use in the crate are
stable, but two of the types aren't, for some reason. However, if we
let type inference do its job via IntoBits, we can avoid having to
import those unstable types and leave all unstable uses for simd
purpose to packed_simd.

For consistency, we use IntoBits everywhere, rather than use FromBits as
well, which would make it necessary to figure out on which specific
configurations FromBits is used.